+++ Makefile.common
@@ -946,8 +946,7 @@ endif
 
 ifeq ($(HAVE_WASAPI), 1)
 	HAVE_MMDEVAPI = 1
-   OBJ += audio/drivers/wasapi.o \
-          audio/common/wasapi.o
+   OBJ += audio/drivers/wasapi.o
    DEFINES += -DHAVE_WASAPI
    LIBS += -lole32 -lksuser
 endif
