+++ audio/common/wasapi.h
@@ -17,24 +17,14 @@
 /**
  * Contains WASAPI-specific support functions that are used
  * by the WASAPI audio and microphone drivers.
- *
  */
 
 #ifndef RETROARCH_COMMON_WASAPI_H
 #define RETROARCH_COMMON_WASAPI_H
 
-#include "mmdevice_common_inline.h"
-#include <boolean.h>
-
 /* Shared buffer size replacement placeholders */
 #define WASAPI_SH_BUFFER_AUDIO_LATENCY 0
 #define WASAPI_SH_BUFFER_DEVICE_PERIOD 32
 #define WASAPI_SH_BUFFER_CLIENT_BUFFER 64
-
-const char *hresult_name(HRESULT hr);
-const char* wasapi_error(DWORD error);
-IMMDevice *wasapi_init_device(const char *id, EDataFlow data_flow);
-IAudioClient *wasapi_init_client(IMMDevice *device, bool *exclusive,
-                                 bool *float_fmt, unsigned *rate, unsigned latency, unsigned channels);
 
 #endif /* RETROARCH_COMMON_WASAPI_H */
